[Ui] Add BindingDraggableBottomSheet
