--- app/src/main/java/com/android/hara/presentation/base/BindingBottomSheetDialog.kt
@@ -1,32 +1,0 @@
-package com.android.hara.presentation.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-abstract class BindingBottomSheetDialog<T : ViewDataBinding>(@LayoutRes private val layoutRes: Int) :
-    BottomSheetDialogFragment() {
-    private var _binding: T? = null
-    val binding get() = requireNotNull(_binding) { "${androidx.appcompat.R.id.content} binding error" }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
-        _binding!!.lifecycleOwner = viewLifecycleOwner
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-
-}
